Updated property display names
