+++ Apps.Systran/Models/Request/TranslateFileRequest.cs
@@ -11,25 +11,25 @@ namespace Apps.Systran.Models.Request
 {
     public class TranslateFileRequest : ITranslateFileInput
     {
-        [Display("Input file")]
+        [Display("File")]
         public FileReference File { get; set; }
 
         [Display("Profile ID")]
         [DataSource(typeof(ProfilesDataHandler))]
         public string? Profile { get; set; }
 
-        [Display("Source language code")]
+        [Display("Source language")]
         [DataSource(typeof(SupportedLanguagesDataHandler))]
         public string Source { get; set; }
 
-        [Display("Target language code")]
+        [Display("Target language")]
         [DataSource(typeof(SupportedLanguagesDataHandler))]
         public string TargetLanguage { get; set; }
 
         [Display("Output file handling", Description = "Determine the format of the output file. The default Blackbird behavior is to convert to XLIFF for future steps."), StaticDataSource(typeof(ProcessFileFormatHandler))]
         public string? OutputFileHandling { get; set; }
 
-        [Display("Translation strategy"), StaticDataSource(typeof(FileTranslationStrategyHandler))]
+        [Display("File translation strategy"), StaticDataSource(typeof(FileTranslationStrategyHandler))]
         public string? FileTranslationStrategy { get; set; } = "blackbird";
     }
 }
